Add KeyError to skippable errors in timeseries
Include KeyError with skippable exceptions for processing
radiance_fid_naive within the context of creating the HIASI-HIRS deviation
plot.  The problem is that filtering such as duplicate removal processing
only happens /after/ the pseudo-fields have been processed.  This means
that pandas indexing in the meantime will fail with InvalidIndexError or
KeyError when duplicates still remain.  The alternative would be to
process the pseudo-fields after the filtering, but that would require
substantial rewriting.